little improvements

+++ src/commonMain/kotlin/eu/vendeli/rethis/types/core/RespVer.kt
@@ -1,7 +1,7 @@
 package eu.vendeli.rethis.types.core
 
 sealed class RespVer(
-    val literal: Short,
+    val literal: Int,
 ) {
     data object V2 : RespVer(2)
     data object V3 : RespVer(3)
